+++ .flake8
@@ -1,3 +1,3 @@
 [flake8]
 max-line-length = 180
-exclude = toro.py, streamparser
+exclude = streamparser
