+++ apertium_apy/modeSearch.py
