+++ apertium_apy/wiki_util.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+
 import logging
 import json
