+++ language_names/README.md
@@ -1,0 +1,7 @@
+Language Names
+==============
+
+To run `scraper.py`, do
+
+    sudo apt-get install libxml2-dev libxslt-dev
+    sudo pip3 install lxml
