+++ requirements.txt
@@ -1,0 +1,7 @@
+flake8==3.5.0
+mccabe==0.6.1
+mypy==0.580
+pycodestyle==2.3.1
+pyflakes==1.6.0
+tornado==4.5.3
+typed-ast==1.1.0
