+++ tests/README
@@ -6,7 +6,7 @@ tools as show at http://wiki.apertium.org/wiki/Debian and then do
     sudo apt-get install apertium-sme-nob apertium-es-en
     mkdir ~/apy-testdata
     cd ~/apy-testdata
-    svn co https://svn.code.sf.net/p/apertium/svn/languages/apertium-nno
+    git clone https://github.com/apertium/apertium-nno
     cd apertium-nno
     ./autogen.sh
     make -j4
