--- tools/README
@@ -1,7 +1,0 @@
-# -*- mode: markdown -*-
-
-To run langNamesScraper.py, do
-
-    sudo apt-get install libxml2-dev libxslt-dev
-	sudo pip3 install lxml
-
