+++ tools/upstart/apertium-apy-gateway.conf
@@ -1,21 +1,21 @@
 description "apertium-apy gateway init script"
-     
+
 start on starting apertium-all
 stop on stopped apertium-all
 respawn
 respawn limit 50 300   # try to respawn 50 times in 300 seconds (5 minutes)
-     
+
 env CONFIG=/etc/default/apertium
 
 script
 	. $CONFIG
-	python3 $APYPATH/gateway.py $SERVERLIST
+	python3 $APYPATH/apertium_apy/gateway.py $SERVERLIST
 end script
 
 pre-start script
 	. $CONFIG
 	touch $LOGFILE
-	echo "`date` $UPSTART_JOB started" >> $LOGFILE 
+	echo "`date` $UPSTART_JOB started" >> $LOGFILE
 end script
 
 post-stop script
