+++ tools/upstart/apertium-apy.conf
@@ -15,7 +15,7 @@ end script
 pre-start script
 	. $CONFIG
 	touch $LOGFILE
-	echo "`date` $UPSTART_JOB started" >> $LOGFILE 
+	echo "`date` $UPSTART_JOB started" >> $LOGFILE
 end script
 
 post-stop script
